services/auth/tests/func/pyproject.toml
[tool.poetry]
name = "auth-component-tests"
version = "0.1.0"
description = ""
authors = ["bbaranov, mkosinov"]

[tool.poetry.dependencies]
python = "^3.12"
pytest-asyncio = "0.12.0"
aiohttp = "3.7.2"
asyncpg = "0.29.0"
pydantic-settings = "2.1.0"
pytest = "^8.1.1"
pycryptodomex = "^3.20.0"
redis = "5.0.3"
asyncio-redis = "^0.16.0"

[tool.isort]
sources = ["./"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode="auto"
